setup de hibernate

+++ myApp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/daos/ClassDao.java
@@ -9,19 +9,19 @@ import java.util.Optional;
 
 public interface ClassDao {
 
-    Optional<Class> get(int id);
+    Optional<Class> get(Long id);
 
-    List<ClassInfo> findClassesByStudentId(int id);
+    List<ClassInfo> findClassesByStudentId(Long id);
 
-    List<ClassInfo> findClassesByTeacherId(int id);
+    List<ClassInfo> findClassesByTeacherId(Long id);
 
-    Class create(int studentId, int teacherId, int level, int subjectId, int price, int status, String message);
+    Class create(Long studentId, Long teacherId, int level, Long subjectId, int price, int status, String message);
 
-    int setStatus(int classId, int status);
+    int setStatus(Long classId, int status);
 
-    int setDeleted(int classId, int deleted);
+    int setDeleted(Long classId, int deleted);
 
-    int setRequest(int classId, String message);
+    int setRequest(Long classId, String message);
 
-    int setReply(int classId, String message);
+    int setReply(Long classId, String message);
 }
