+++ myApp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/daos/ImageDao.java
@@ -5,7 +5,7 @@ import ar.edu.itba.paw.models.Image;
 import java.util.Optional;
 
 public interface ImageDao {
-    Image createOrUpdate(int uid, byte[] image);
+    Image createOrUpdate(Long uid, byte[] image);
 
-    Optional<Image> findImageById(int userId);
+    Optional<Image> findImageById(Long userId);
 }
