+++ myApp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/daos/SubjectDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface SubjectDao {
-    Optional<Subject> findById(int id);
+    Optional<Subject> findById(Long id);
 
     Subject create (String subject);
