+++ myApp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/ImageService.java
@@ -5,8 +5,8 @@ import ar.edu.itba.paw.models.Image;
 import java.util.Optional;
 
 public interface ImageService {
-    Image createOrUpdate(int uid, byte[] image);
+    Image createOrUpdate(Long uid, byte[] image);
 
-    Optional<Image> findImageById(int userId);
+    Optional<Image> findImageById(Long userId);
 
 }
