+++ myApp/persistence/src/main/java/ar/edu/itba/paw/persistence/ImageDaoJdbc.java
@@ -18,7 +18,7 @@ public class ImageDaoJdbc implements ImageDao {
     private final JdbcTemplate jdbcTemplate;
     private SimpleJdbcInsert jdbcInsert;
     private final static RowMapper<Image> ROW_MAPPER = (rs, rowNum) ->
-            new Image(rs.getInt("userid"), rs.getBytes("image"));
+            new Image(rs.getLong("userid"), rs.getBytes("image"));
 
     @Autowired
     public ImageDaoJdbc(final DataSource ds) {
@@ -28,7 +28,7 @@ public class ImageDaoJdbc implements ImageDao {
     }
 
     @Override
-    public Image createOrUpdate(int uid, byte[] image) {
+    public Image createOrUpdate(Long uid, byte[] image) {
         jdbcTemplate.update("insert into images values (?,?)\n" +
                 " on conflict on constraint images_pkey\n" +
                 "     do update set image = excluded.image;",uid, image);
@@ -36,7 +36,7 @@ public class ImageDaoJdbc implements ImageDao {
     }
 
     @Override
-    public Optional<Image> findImageById(int userId) {
+    public Optional<Image> findImageById(Long userId) {
         List<Image> list = jdbcTemplate.query("SELECT * FROM images WHERE userid = ?", new Object[] { userId }, ROW_MAPPER);
         return list.isEmpty() ? Optional.empty() : Optional.of(list.get(0));
     }
